splitting files into 3
