+++ test/test_application_1.py
@@ -16,6 +16,7 @@ from test.pages import (  # pylint: disable=wrong-import-order
     get_browser,
 )
 from test.test_utility_functions import (
+    async_test,
     flush_database,
     flush_redis,
     client_session,
@@ -42,15 +43,6 @@ from test.test_utility_functions import (
     setup_elasticsearch_indexes,
     b64_decode,
 )
-
-
-def async_test(func):
-    def wrapper(*args, **kwargs):
-        future = func(*args, **kwargs)
-        loop = asyncio.get_event_loop()
-        loop.run_until_complete(future)
-
-    return wrapper
 
 
 class TestApplication(unittest.TestCase):
@@ -1073,40 +1065,3 @@ class TestApplication(unittest.TestCase):
         self.assertEqual(
             payload["authorised_hosts"], ["mlflow--data-science--internal.dataworkspace.test:8000"]
         )
-
-    @async_test
-    async def test_application_redirects_to_sso_again_if_token_expired(self):
-        await flush_database()
-        await flush_redis()
-
-        session, cleanup_session = client_session()
-        self.add_async_cleanup(cleanup_session)
-
-        cleanup_application = await create_application()
-        self.add_async_cleanup(cleanup_application)
-
-        is_logged_in = True
-        codes = iter(["some-code", "some-other-code"])
-        tokens = iter(["token-1", "token-2"])
-        auth_to_me = {
-            # No token-1
-            "Bearer token-2": {
-                "email": "test@test.com",
-                "contact_email": "test@test.com",
-                "related_emails": [],
-                "first_name": "Peter",
-                "last_name": "Piper",
-                "user_id": "7f93c2c7-bc32-43f3-87dc-40d0b8fb2cd2",
-            }
-        }
-        sso_cleanup, number_of_times_at_sso = await create_sso(
-            is_logged_in, codes, tokens, auth_to_me
-        )
-        self.add_async_cleanup(sso_cleanup)
-
-        await until_succeeds("http://dataworkspace.test:8000/healthcheck")
-
-        # Make a request to the home page
-        async with session.request("GET", "http://dataworkspace.test:8000/") as response:
-            self.assertEqual(number_of_times_at_sso(), 2)
-            self.assertEqual(200, response.status)
