+++ test/test_utility_functions.py
@@ -13,6 +13,15 @@ from elasticsearch import AsyncElasticsearch, helpers
 from lxml import html
 
 
+def async_test(func):
+    def wrapper(*args, **kwargs):
+        future = func(*args, **kwargs)
+        loop = asyncio.get_event_loop()
+        loop.run_until_complete(future)
+
+    return wrapper
+
+
 def client_session():
     session = aiohttp.ClientSession()
